v 0.11.0
Major overhaul
Switched up the pattern a little bit
Created "drawables", which are the javascript-side components for a
given shader program, and all models associated with it
Created "entities", which are simple instances of various models,
with some sensible defaults.  Entities rely on an AssetManager,
and that all resources required are already preloaded.

The base drawable is found at src/drawable.js
Drawables can be found in src/drawables/
Entity base can be found in src/entity.js
Entities themselves can be found in src/entities/

Added basic entities for:
Portal
Shield Effect
.. all inventory items