Adding resonator links Various upgrades to link meshes. Various upgrades to meshes
Entities now accept an instance of an engine, so they can properly wire up/teardown all their component drawables.
This is still very much early work on this, but I don't expect the requirement of the engine to change.  Perhaps
the engine itself will provide constructors for entities, I dunno.