Big pile of updates in order to support rebuilding ingress-models.info
So... a while ago Google Cloud decided to end-of-life the instance the site was running on.  By the time I got around to trying to deploy it on a new instance, I discovered that I had been bitten by a pretty severe case of *toolchain rot*.  None of the site would build, and the machine that had the scripts I used in the past was no longer in service... So I decided to upgrade and modernize things a bit, and just start over.

This necessitated a handful of changes here--some necessary, some more just for quality of life when developing on a newer toolchain

- Switch to `webpack` from `grunt`
- Move some files to be in more "standard" locations
- Add a `Caddyfile` for easy hosting for local development
- Switch to `yarn`
- Update "Inventory Entities" to actually just be nested `Drawable`s, as I had intended (one TODO down!)
- Clean up a couple of the demos to work better
- Add a secondary webpack configuration for a "bundled" build (i.e. minified and all)
- Update `eslint` rules
- Remove the dependency on `es6-promises`